Adding wireframe files.
This addition will include the basic setup files for he webpage to include wireframing to allow it to be used for all sized desktops.

The icon section covers approximately 33% of the right-hand side of the screen with the other widgets covering the other 67%. Of the right portion the 3D map will cover the top half and a sound/info/alternate widget will split the bottom half with a notepad/codepad widget.